+++ src/nodes/EditLevelLayer.hpp
@@ -2,20 +2,15 @@
 #pragma once
 
 #include <Geode/Geode.hpp>
-#include <Geode/modify/EditLevelLayer.hpp>
 #include "../Utils.hpp"
 #include "../Macros.hpp"
+#include <alphalaneous.alphas_geode_utils/include/NodeModding.h>
 
 using namespace geode::prelude;
 
-class $modify(MyEditLevelLayer, EditLevelLayer) {
+class $nodeModify(MyEditLevelLayer, EditLevelLayer) {
 
-    static void onModify(auto& self) {
-        HOOK_LATEST("EditLevelLayer::init");
-    }
-
-    bool init(GJGameLevel* p0) {
-        if (!EditLevelLayer::init(p0)) return false;
+    void modify() {
         if (UIModding::get()->doModify) {
             if (CCScale9Sprite* bg = typeinfo_cast<CCScale9Sprite*>(this->getChildByID("description-background"))) {
                 Utils::setColorIfExists(bg, "edit-description-bg");
@@ -24,6 +19,5 @@ class $modify(MyEditLevelLayer, EditLevelLayer) {
                 Utils::setColorIfExists(bg, "edit-name-bg");
             }
         }
-        return true;
     }
 };