+++ src/nodes/LevelSearchLayer.hpp
@@ -1,19 +1,14 @@
 #pragma once
 
 #include <Geode/Geode.hpp>
-#include <Geode/modify/LevelSearchLayer.hpp>
 #include "../Macros.hpp"
+#include <alphalaneous.alphas_geode_utils/include/NodeModding.h>
+#include "../UIModding.hpp"
 
 using namespace geode::prelude;
 
-class $modify(MyLevelSearchLayer, LevelSearchLayer) {
-
-    static void onModify(auto& self) {
-        HOOK_LATEST("LevelSearchLayer::init");
-    }
-
-    bool init(int p0) {
-        if (!LevelSearchLayer::init(p0)) return false;
+class $nodeModify(MyLevelSearchLayer, LevelSearchLayer) {
+    void modify() {
         if (UIModding::get()->doModify) {
             UIModding::get()->updateColors(this, "level-search-bg");
             UIModding::get()->updateColors(this, "level-search-bar-bg");
@@ -21,6 +16,5 @@ class $modify(MyLevelSearchLayer, LevelSearchLayer) {
             UIModding::get()->updateColors(this, "difficulty-filters-bg");
             UIModding::get()->updateColors(this, "length-filters-bg");
         }
-        return true;
     }
 };