+++ src/nodes/MenuLayer.hpp
@@ -1,23 +1,16 @@
 #pragma once
 
 #include <Geode/Geode.hpp>
-#include <Geode/modify/MenuLayer.hpp>
 #include "../Callbacks.hpp"
 #include "../Macros.hpp"
 #include "../UIModding.hpp"
+#include <alphalaneous.alphas_geode_utils/include/NodeModding.h>
 
 using namespace geode::prelude;
 
-class $modify(MyMenuLayer, MenuLayer) {
+class $nodeModify(MyMenuLayer, MenuLayer) {
 
-    static void onModify(auto& self) {
-        HOOK_LATEST("MenuLayer::init");
-    }
-
-    bool init() {
-        UIModding::get()->finishedLoad = true;
-        if (!MenuLayer::init()) return false;
-
+    void modify() {
         UIModding::get()->doModify = Mod::get()->getSettingValue<bool>("ui-modifications");
         Callbacks::get()->generateAll();
 
@@ -26,7 +19,5 @@ class $modify(MyMenuLayer, MenuLayer) {
                 UIModding::get()->startFileListeners();
             }
         }
-
-        return true;
     }
 };